Use source timestamp when generating manpages
When generating the man page, the timestamp of the generated file is
taken from todays date, which obviously change every day.  This make
repeated builds change for no sensible reason.  Because of this, I
suggest to change the manual page generation code to use the time of the
last change to the manual.py file instead.