Add a --domaindir option to support dynamic --domain changes
The new --domaindir option allows for a directory of domains to be
specified when calling the script. This option will only work if
the instance of the script is a frontend, i.e. if --isfrontend was
specified when starting the instance. In this case, the frontend
instance will look in the specified directory for --domain=
specifications and update its internal dictionary. Domains can be
added and removed dynamically during the life of the instance.

Backends that are already connected will not be closed by the
frontend on having their --domain line removed from the domain
directory, but will be rejected on future reconnections.

It is expected that the path specified through the --domaindir
option only contains files specifying --domain options. The
internal backend dictionary of the frontend instance will only be
updated on the next attempted connection from a backend instance,
and if the timestamp of the domain directory has been updated.
The recommended way to update the domain directory is therefore
either to use the "touch" Linux utility, or to only change the
directory through removal and creation of new files.

Since this option is implemented using the existing
ConfigureFromFile function, it may allow dynamical changes to any
configuration parameter of the instance, but only domains have
been tested and are supported through the above-specified usage
recommendations.