Only update epoll mask if it changes
In every epoll iteration, epoll_modify is called for all connections to
set the mask. However, in most cases, the mask doesn't change. With
thousands of connections, updating the mask takes a lot of time,
completely offsetting the advantage of epoll over select.

To fix this, store the old mask in the selectable object, and only do
epoll_modify if the mask changes.