Stop accepting when there are too many queued connections
The introduction of the fastping helper in a separate thread creates a
problem for frontends that handle a large number of kites. When the
frontend server starts, all of those kites will start connecting to the
frontend, many of them at the same time. This creates a very long queue
of connections that need active attention. Adding all of them to the
fastping and connections queues slows things down dramatically for
everybody, to the point that most connections time out. This is
especially relevant when an external authdomain tool is used, because it
can significantly slow down the handling of a new connection.

To work around this issue, limit the number of connections that are
allowed to be in the fastping queue or in the new state. When more than
10 connections are in this state, stop accepting new connections. Or
rather, give the connections in the new state more opportunity by
accepting only a single new connection per epoll iteration.

Tested by starting 400 pagekitec backends simultaneously, and with an
authdomain script that sleeps 200ms. Without this change, about 120
connections "survive" after 2 minutes. With this change, the number
increases to 250.